+++ lib/mojito.ex
@@ -11,13 +11,95 @@ defmodule Mojito do
   one-off requests as well, using the same process-less architecture as
   Mint.
 
+  ## Quickstart
+
+      {:ok, response} = Mojito.request(method: :get, url: "https://github.com")
+
+  ## Why Mojito?
+
+  Mojito exists to fill an as-yet-vacant niche in Erlang and Elixir HTTP
+  clients, combining the following design goals:
+
+  * _Little or no configuration needed._  Use Mojito to make requests to as
+    many different destinations as you like, without thinking about
+    starting or selecting connection pools.  Other clients like
+    [Hackney](https://github.com/benoitc/hackney)
+    (and [HTTPoison](https://github.com/edgurgel/httpoison)),
+    [Ibrowse](https://github.com/cmullaparthi/ibrowse) (and
+    [HTTPotion](https://github.com/myfreeweb/httpotion)), and
+    Erlang's built-in httpc offer this feature.
+
+  * _Connection pools used only for a single destination._  Using a pool
+    for making requests against multiple destinations is less than ideal,
+    as many of the connections need to be reset before use.  Mojito assigns
+    requests to the correct pools transparently to the user.  Other clients,
+    such as [Buoy](https://github.com/lpgauth/buoy), Hackney/HTTPoison, and
+    [Gun](https://github.com/ninenines/gun), force the user to handle this
+    themselves, which is often inconvenient if the full set of HTTP
+    destinations is not known at compile time.
+
+  * _Redundant pools to reduce GenServer-related bottlenecks._  Mojito can
+    serve requests to the same destination from more than one connection
+    pool, and those pools can be selected by round-robin at runtime in order
+    to minimize resource contention in the Erlang VM.  No other HTTP client
+    for Erlang or Elixir supports this.
+
+  * _Ludicrous speed._  Mojito meets or exceeds the performance of
+    every other HTTP client for Erlang or Elixir.
+    [Check out the benchmarks.](https://github.com/appcues/mojito/blob/master/BENCHMARKS.md)
+
   ## Installation
 
   Add `mojito` to your deps in `mix.exs`:
 
       {:mojito, "~> 0.3.0"}
 
-  ## Common usage
+  ## Configuration
+
+  The following `config.exs` config parameters are supported:
+
+  * `:timeout` (milliseconds, default 5000) -- Default request timeout.
+  * `:transport_opts` (`t:Keyword.t`, default `[]`) -- Options to pass to
+    the `:gen_tcp` or `:ssl` modules.  Commonly used to make HTTPS requests
+    with self-signed TLS server certificates; see below for details.
+  * `:pool_opts` (`t:pool_opts`, default `[]`) -- Configuration options
+    for connection pools.
+
+  The following `:pool_opts` options are supported:
+
+  * `:size` (integer) sets the number of steady-state connections per pool.
+    Default is 5.
+  * `:max_overflow` (integer) sets the number of additional connections
+    per pool, opened under conditions of heavy load.
+    Default is 10.
+  * `:pools` (integer) sets the maximum number of pools to open for a
+    single destination host and port (not the maximum number of total
+    pools to open).  Default is 5.
+  * `:strategy` is either `:lifo` or `:fifo`, and selects which connection
+    should be checked out of a single pool.  Default is `:lifo`.
+  * `:destinations` (keyword list of `t:pool_opts`) allows these parameters
+    to be set for individual `:"host:port"` destinations.
+
+  For example:
+
+      use Mix.Config
+
+      config :mojito,
+        timeout: 2500,
+        pool_opts: [
+          size: 10,
+          destinations: [
+            "example.com:443": [
+              size: 20,
+              max_overflow: 20,
+              pools: 10
+            ]
+          ]
+        ]
+
+  Certain configs can be overridden with each request.  See `request/1`.
+
+  ## Usage
 
   Make requests with `Mojito.request/1` or `Mojito.request/5`:
 
@@ -41,9 +123,8 @@ defmodule Mojito do
 
   By default, Mojito will use a connection pool for requests, automatically
   handling the creation and reuse of pools.  If this is not desired,
-  specify the `pool: false` option with a request to perform a one-off request,
-  or `pool: pool_name_or_pid` to use a specific user-created `Mojito.Pool.Single`
-  pool.  See the documentation for `request/1` for more details.
+  specify the `pool: false` option with a request to perform a one-off request.
+  See the documentation for `request/1` for more details.
 
   ## Self-signed SSL/TLS certificates
 
@@ -56,6 +137,33 @@ defmodule Mojito do
 
       >>>> Mojito.request(method: :get, url: "https://localhost:8443/", opts: [transport_opts: [verify: :verify_none]])
       {:ok, %Mojito.Response{ ... }}
+
+  ## Changelog
+
+  See the [CHANGELOG.md](https://github.com/appcues/mojito/blob/master/CHANGELOG.md).
+
+  ## Contributing
+
+  Thanks for considering contributing to this project, and to the free
+  software ecosystem at large!
+
+  Interested in contributing a bug report?  Terrific!  Please open a [GitHub
+  issue](https://github.com/appcues/mojito/issues) and include as much detail
+  as you can.  If you have a solution, even better -- please open a pull
+  request with a clear description and tests.
+
+  Have a feature idea?  Excellent!  Please open a [GitHub
+  issue](https://github.com/appcues/mojito/issues) for discussion.
+
+  Want to implement an issue that's been discussed?  Fantastic!  Please
+  open a [GitHub pull request](https://github.com/appcues/mojito/pulls)
+  and write a clear description of the patch.
+  We'll merge your PR a lot sooner if it is well-documented and fully
+  tested.
+
+  Contributors and contributions are listed in the
+  [changelog](https://github.com/appcues/mojito/blob/master/CHANGELOG.md).
+  Heartfelt thanks to everyone who's helped make Mojito better.
 
   ## Authorship and License
 
@@ -98,6 +206,14 @@ defmodule Mojito do
           message: String.t() | nil,
         }
 
+  @type pool_opts :: [pool_opt | {:destinations, [{atom, pool_opts}]}]
+
+  @type pool_opt ::
+          {:size, pos_integer}
+          | {:max_overflow, non_neg_integer}
+          | {:pools, pos_integer}
+          | {:strategy, :lifo | :fifo}
+
   @doc ~S"""
   Performs an HTTP request and returns the response.
 
@@ -132,10 +248,6 @@ defmodule Mojito do
   in this case, it is recommended to wrap `request/1` in `Task.async/1`,
   or use one of the pooled request modes.
 
-  `pool: <pid or name>` can be provided to run the request against a
-  manually-created `Mojito.Pool.Single` pool using
-  `Mojito.Pool.Single.request/2`.
-
   Options:
 
   * `:pool` - See above.
@@ -155,8 +267,8 @@ defmodule Mojito do
       end
 
       ## Retry connection-closed errors once
-      case request_fn.() do
-        {:error, %{message: {:error, _, %{reason: :closed}}}} -> request_fn.()
+      case request_fn.() |> Mojito.Utils.wrap_return_value do
+        {:error, %{reason: %{reason: :closed}}} -> request_fn.()
         other -> other
       end
     end