+++ lib/mojito/utils.ex
@@ -15,6 +15,8 @@ defmodule Mojito.Utils do
       {:ok, _} -> rv
       {:error, %Error{}} -> rv
       {:error, e} -> {:error, %Error{reason: e}}
+      {:error, {:error, e}} -> {:error, %Error{reason: e}}
+      {:error, _mint_conn, error} -> {:error, %Error{reason: error}}
       other -> {:error, %Error{reason: :unknown, message: other}}
     end
   end